requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
effdet==0.3.0
ensemble-boxes==1.0.9
fastcore==1.5.27
filelock==3.8.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.23.0
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.4.2
llvmlite==0.39.1
matplotlib==3.6.2
multidict==6.0.3
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
omegaconf==2.3.0
opencv-python-headless==4.6.0.66
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.1
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.8.6
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
